Update to 2.3.0, based on a patch by Lokesh Mandvekar <lsm5@fedoraproject.org> on pkgsrc-users.
Changes:
This release adds the ability to move/resize floating windows beyond
region boundaries. It adds 'soft boundary' behavior to region
boundaries. When moving a window past the region boundary, the
window will 'snap' to the region boundary if it is less than
boundary_width distance beyond the edge. A new boundary_width
configuration option has been added. The 'soft boundary' behavior
can be disabled by setting this option to 0. The ability to set
tile_gap to negative values has been added, which makes it possible
for tiled windows to overlap. Set this to the opposite of border_width
to collapse borders.