Add PHP 5.5.1 as lang/php55 version 5.5.1 package.
This is new stable release of PHP.  Please refer UPGRADING file for
changes and updating.


PHP is an HTML-embedded scripting language. It is modular, with
some object-oriented features. Much of its syntax is borrowed from
C, Java and Perl with a couple of unique PHP-specific features
thrown in.  The language is designed to allow web developers to
write dynamically generated pages quickly.

This package provices PHP version 5.5.x.