Update to 1.6.1
Upstream changes:
Changes in version 1.6-1

write.matrix.csr() now accepts a fac argument similar to read.matrix.csr(), writing factor levels instead of the numeric codes.

naiveBayes() uses a numerically more stable formula for calculating the a-posterior probabilities.

predict.naiveBayes() now accepts data with predictors in an order different from the training data, and also ignores variables not in the model (especially the response variable).

svm() checks whether parameters which are passed to the C-code are set to NULL to avoid segfaults.