Switch to textproc/mdocml as the tool replacement provider for 'nroff'.
The nroff tool is primarily used to generate catman pages, however there
were circular dependency issues with using groff as the provider, as it
has a large number of dependencies, some of which depend on nroff.

mdocml is much smaller, has fewer dependencies, is BSD licensed, and is
just as capable in nroff mode.