Update to Asterisk 11.6.1: this is a security fix update to fix AST-2013-006 and AST-2013-007, and a minor bug fix update.
pkgsrc change: disable SRTP on NetBSD as it doesn't link

---- 11.6.1 ----

The Asterisk Development Team has announced security releases for Certified
Asterisk 1.8.15, 11.2, and Asterisk 1.8, 10, and 11. The available security
releases are released as versions 1.8.15-cert4, 11.2-cert3, 1.8.24.1, 10.12.4,
10.12.4-digiumphones, and 11.6.1.

The release of these versions resolve the following issues:

* A buffer overflow when receiving odd length 16 bit messages in app_sms. An
  infinite loop could occur which would overwrite memory when a message is
  received into the unpacksms16() function and the length of the message is an
  odd number of bytes.

* Prevent permissions escalation in the Asterisk Manager Interface. Asterisk
  now marks certain individual dialplan functions as 'dangerous', which will
  inhibit their execution from external sources.

  A 'dangerous' function is one which results in a privilege escalation. For
  example, if one were to read the channel variable SHELL(rm -rf /) Bad
  Things(TM) could happen; even if the external source has only read
  permissions.

  Execution from external sources may be enabled by setting 'live_dangerously'
  to 'yes' in the [options] section of asterisk.conf. Although doing so is not
  recommended.

These issues and their resolutions are described in the security advisories.

For more information about the details of these vulnerabilities, please read
security advisories AST-2013-006 and AST-2013-007, which were
released at the same time as this announcement.

For a full list of changes in the current releases, please see the ChangeLogs:

http://downloads.asterisk.org/pub/telephony/asterisk/releases/ChangeLog-11.6.1

The security advisories are available at:

 * http://downloads.asterisk.org/pub/security/AST-2013-006.pdf
 * http://downloads.asterisk.org/pub/security/AST-2013-007.pdf

Thank you for your continued support of Asterisk!

----- 11.6.0 -----

The Asterisk Development Team has announced the release of Asterisk 11.6.0.

The release of Asterisk 11.6.0 resolves several issues reported by the
community and would have not been possible without your participation.
Thank you!

The following is a sample of the issues resolved in this release:

* --- Confbridge: empty conference not being torn down
  (Closes issue ASTERISK-21859. Reported by Chris Gentle)

* --- Let Queue wrap up time influence member availability
  (Closes issue ASTERISK-22189. Reported by Tony Lewis)

* --- Fix a longstanding issue with MFC-R2 configuration that
      prevented users
  (Closes issue ASTERISK-21117. Reported by Rafael Angulo)

* --- chan_iax2: Fix saving the wrong expiry time in astdb.
  (Closes issue ASTERISK-22504. Reported by Stefan Wachtler)

* --- Fix segfault for certain invalid WebSocket input.
  (Closes issue ASTERISK-21825. Reported by Alfred Farrugia)

For a full list of changes in this release, please see the ChangeLog:

http://downloads.asterisk.org/pub/telephony/asterisk/ChangeLog-11.6.0

Thank you for your continued support of Asterisk!